feat: rename substrate to internal
