+++ internal/dao/chainBlock.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 	"github.com/go-kratos/kratos/pkg/cache/redis"
 	"github.com/itering/subscan/internal/model"
-	"github.com/itering/subscan/libs/substrate"
-	"github.com/itering/subscan/libs/substrate/rpc"
+	"github.com/itering/subscan/internal/substrate"
+	"github.com/itering/subscan/internal/substrate/rpc"
 	"github.com/itering/subscan/util"
 	"github.com/itering/subscan/util/ss58"
 	"sort"
