+++ internal/dao/chainExtrinsic.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"encoding/json"
 	"github.com/itering/subscan/internal/model"
-	"github.com/itering/subscan/libs/substrate"
+	"github.com/itering/subscan/internal/substrate"
 	"github.com/itering/subscan/util"
 	"github.com/itering/subscan/util/ss58"
 	"github.com/shopspring/decimal"
