+++ internal/dao/chainLog.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"fmt"
 	"github.com/itering/subscan/internal/model"
-	"github.com/itering/subscan/libs/substrate/storage"
+	"github.com/itering/subscan/internal/substrate/storage"
 	"github.com/itering/subscan/util"
 	"strings"
 )
