+++ internal/dao/extrinsicError.go
@@ -2,7 +2,7 @@ package dao
 
 import (
 	"github.com/itering/subscan/internal/model"
-	"github.com/itering/subscan/libs/substrate"
+	"github.com/itering/subscan/internal/substrate"
 	"github.com/itering/subscan/util"
 	"strings"
 )
