+++ internal/server/http/scan.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	bm "github.com/go-kratos/kratos/pkg/net/http/blademaster"
 	"github.com/go-kratos/kratos/pkg/net/http/blademaster/binding"
-	"github.com/itering/subscan/libs/substrate"
+	"github.com/itering/subscan/internal/substrate"
 	"github.com/itering/subscan/util"
 	"github.com/itering/subscan/util/ss58"
 	"regexp"
