+++ internal/service/chainLog.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"github.com/itering/subscan/internal/dao"
 	"github.com/itering/subscan/internal/model"
-	"github.com/itering/subscan/libs/substrate"
-	"github.com/itering/subscan/libs/substrate/storage"
+	"github.com/itering/subscan/internal/substrate"
+	"github.com/itering/subscan/internal/substrate/storage"
 	"github.com/itering/subscan/util"
 	"strings"
 )
