+++ internal/service/substrate.go
@@ -7,9 +7,9 @@ import (
 	"fmt"
 	"github.com/freehere107/go-workers"
 	"github.com/go-kratos/kratos/pkg/log"
-	"github.com/itering/subscan/libs/substrate"
-	"github.com/itering/subscan/libs/substrate/rpc"
-	"github.com/itering/subscan/libs/substrate/websocket"
+	"github.com/itering/subscan/internal/substrate"
+	"github.com/itering/subscan/internal/substrate/rpc"
+	"github.com/itering/subscan/internal/substrate/websocket"
 	"github.com/itering/subscan/util"
 	"github.com/panjf2000/ants/v2"
 	"sync"
