+++ internal/substrate/event.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	scalecodec "github.com/freehere107/go-scale-codec"
 	"github.com/freehere107/go-scale-codec/types"
-	"github.com/itering/subscan/libs/substrate/metadata"
+	"github.com/itering/subscan/internal/substrate/metadata"
 	"github.com/itering/subscan/util"
 )
