+++ internal/substrate/hasher/hasher.go
