+++ internal/substrate/log.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/freehere107/go-scale-codec/types"
-	"github.com/itering/subscan/libs/substrate/storage"
+	"github.com/itering/subscan/internal/substrate/storage"
 	"github.com/itering/subscan/util"
 )
