+++ internal/substrate/storageKey/storageKey.go
@@ -3,8 +3,8 @@ package storageKey
 import (
 	"fmt"
 	"github.com/freehere107/go-scale-codec/types"
-	"github.com/itering/subscan/libs/substrate/hasher"
-	"github.com/itering/subscan/libs/substrate/metadata"
+	"github.com/itering/subscan/internal/substrate/hasher"
+	"github.com/itering/subscan/internal/substrate/metadata"
 	"github.com/itering/subscan/util"
 )
